// platform-api/src/worker/enum/worker-type.enum.ts
export enum WorkerType {
  NODE_DEFAULT = 'node_default',
  NODE_NESTJS = 'node_nestjs',
  NODE_REACTJS_CYPRESS = 'node_reactjs_cypress',
  NODE_GRPCJS = 'node_grpcjs',
  NODE_NEXTJS_CYPRESS = 'node_nextjs_cypress',
  NODE_DEFAULT_POSTGRESQL = 'node_default_postgresql',
  NODE_NESTJS_POSTGRESQL = 'node_nestjs_postgresql',
  NODE_TERAORM = 'node_teraorm',
  PYTHON_DEFAULT = 'python_default',
}